Add notes to the patient page
