+++ public/locale/en.json
@@ -1675,26 +1675,6 @@
   "encounter_locations": "Encounter Locations",
   "encounter_manage_organization_description": "Add or remove organizations from this encouter",
   "encounter_marked_as_complete": "Encounter Completed",
-  "encounter_notes__all_discussions": "All Discussions",
-  "encounter_notes__all_discussions_description": "View and manage encounternotes discussion threads",
-  "encounter_notes__be_first_to_send": "Be the first to send a message",
-  "encounter_notes__choose_template": "Choose a template or enter a custom title",
-  "encounter_notes__create_discussion": "Create a new discussion thread to organize your conversation topics.",
-  "encounter_notes__discussions": "Discussions",
-  "encounter_notes__enter_discussion_title": "Enter discussion title...",
-  "encounter_notes__failed_create_thread": "Failed to create thread",
-  "encounter_notes__failed_send_message": "Failed to send message",
-  "encounter_notes__inactive_encounter": "You cannot start a new discussion for this encounter as it is {{encounterStatus}}.",
-  "encounter_notes__new": "New",
-  "encounter_notes__no_discussions": "No discussions yet",
-  "encounter_notes__no_unused_threads": "Please enter a custom title for thread",
-  "encounter_notes__select_create_thread": "Select or create a thread to start messaging",
-  "encounter_notes__start_conversation": "Start the Conversation",
-  "encounter_notes__start_new_discussion": "Start New Discussion",
-  "encounter_notes__thread_created": "Thread created successfully",
-  "encounter_notes__type_message": "Type your message...",
-  "encounter_notes__welcome": "Welcome to Discussions",
-  "encounter_notes__welcome_description": "Start a new discussion or select an existing thread to begin messaging",
   "encounter_priority__ASAP": "ASAP",
   "encounter_priority__as_needed": "As needed",
   "encounter_priority__asap": "ASAP",
@@ -2849,6 +2829,26 @@
   "not_started": "Not started",
   "note": "Note",
   "notes": "Notes",
+  "notes__all_discussions": "All Discussions",
+  "notes__all_discussions_description": "View and manage encounternotes discussion threads",
+  "notes__be_first_to_send": "Be the first to send a message",
+  "notes__choose_template": "Choose a template or enter a custom title",
+  "notes__create_discussion": "Create a new discussion thread to organize your conversation topics.",
+  "notes__discussions": "Discussions",
+  "notes__enter_discussion_title": "Enter discussion title...",
+  "notes__failed_create_thread": "Failed to create thread",
+  "notes__failed_send_message": "Failed to send message",
+  "notes__inactive_encounter": "You cannot start a new discussion for this encounter as it is {{encounterStatus}}.",
+  "notes__new": "New",
+  "notes__no_discussions": "No discussions yet",
+  "notes__no_unused_threads": "Please enter a custom title for thread",
+  "notes__select_create_thread": "Select or create a thread to start messaging",
+  "notes__start_conversation": "Start the Conversation",
+  "notes__start_new_discussion": "Start New Discussion",
+  "notes__thread_created": "Thread created successfully",
+  "notes__type_message": "Type your message...",
+  "notes__welcome": "Welcome to Discussions",
+  "notes__welcome_description": "Start a new discussion or select an existing thread to begin messaging",
   "notes_placeholder": "Type your Notes",
   "notice_board": "Notice Board",
   "notification_cancelled": "Notification cancelled",
