+++ src/components/Patient/PatientDetailsTab/PatientNotes.tsx
@@ -1,0 +1,14 @@
+import { NoteManager } from "@/components/Notes/NoteManager";
+
+import { PatientProps } from ".";
+
+export const PatientNotesTab = (props: PatientProps) => {
+  return (
+    <NoteManager
+      canAccess={true}
+      canWrite={true}
+      patientId={props.patientData.id}
+      encounterId={undefined}
+    />
+  );
+};
